Propagation (Front/Back)
Rewrote forward-prop to A) actually be right and B) use a lispy seeming solution

Wrote back-prop based on Forward-prop. It looks like the error evaluation is meant to happen in the higher functions net-build and simple-generalization, so I didn't do any of that in these two. It would be fairly trivial to add if we decided it should be there, but I don't think it is.